fix erros
