+++ package.json
@@ -44,6 +44,7 @@
   "devDependencies": {
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.1.1",
+    "@types/react-redux": "^7.1.24",
     "identity-obj-proxy": "^3.0.0",
     "jest": "^27.5.1",
     "ts-jest": "^27.1.4"
