+++ src/domain/nav-item.ts
@@ -1,0 +1,5 @@
+export interface INavItem {
+  id: string;
+  title: string;
+  to: string;
+}
